% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covC}
\alias{covC}
\title{calculate covariance matrix}
\usage{
covC(X)
}
\arguments{
\item{X}{design matrix}
}
\value{
covariance matrix
}
\description{
calculate covariance matrix
}
\keyword{internal}
