% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covCdaC}
\alias{covCdaC}
\title{Optimize a linear regression model by coordinate descent algorithm using a covariance matrix}
\usage{
covCdaC(
  Gamma_,
  gamma_,
  lambda_,
  init_beta,
  beta_tilde_,
  penalty_factor_lasso_,
  penalty_factor_trlasso_,
  alpha = 1,
  maxit = 10000,
  eps = 1e-04,
  strong = TRUE
)
}
\arguments{
\item{init_beta}{initial values of beta}

\item{alpha}{alpha}

\item{maxit}{max iteration}

\item{eps}{convergence threshold for optimization}

\item{strong}{whether use strong screening or not}

\item{Gamma}{covariance matrix of explanatory variables}

\item{gamma}{covariance vector of explanatory and objective variables}

\item{lambda}{lambda sequence}

\item{beta_tilde}{initial estimates (source parameters) of beta}
}
\value{
standardized beta
}
\description{
Optimize a linear regression model by coordinate descent algorithm using a covariance matrix
}
\keyword{internal}
