% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cv.trlasso.mse}
\alias{cv.trlasso.mse}
\title{Compute averaged error of predicted target variable.}
\usage{
cv.trlasso.mse(family, actual, pred)
}
\arguments{
\item{family}{family of regression: "gaussian"}

\item{actual}{matrix of actual target variable}

\item{pred}{matrix of predicted target variable}
}
\description{
Compute averaged error of predicted target variable.
}
