% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validator.R
\name{cv.trlasso}
\alias{cv.trlasso}
\title{Fit a model using a design matrix with cross validation}
\usage{
cv.trlasso(
  X,
  y,
  beta_tilde = NULL,
  family = "gaussian",
  nfolds = 10,
  lambda.min.ratio = 0.01,
  nlambda = 100,
  lambda = NULL,
  alpha = 1,
  foldid = NULL,
  unit = "sample",
  seed = NULL,
  eps = 1e-04,
  verbose = FALSE,
  standardize = TRUE,
  penalty.factor.lasso = NULL,
  penalty.factor.trlasso = NULL,
  penalty.factor.thresh = 0.01,
  ...
)
}
\arguments{
\item{X}{matrix of explanatory variables}

\item{y}{vector of objective variable}

\item{beta_tilde}{vector of initial parameters}

\item{family}{family of regression: "gaussian" is only supported}

\item{nfolds}{the number of folds (ignored if foldid is specified)}

\item{lambda.min.ratio}{ratio of max lambda and min lambda (ignored if lambda is specified)}

\item{nlambda}{the number of lambda (ignored if lambda is specified)}

\item{lambda}{lambda sequence}

\item{alpha}{alpha}

\item{foldid}{vector indicating id of fold for each sample}

\item{unit}{unit for cross validation error: "sample" (default) or "fold"}

\item{seed}{random seed of cross validation}

\item{eps}{convergence threshold (default 1e-4)}

\item{verbose}{whether output verbose warnings and messages (default FALSE)}

\item{standardize}{standardize or not (experimental, not safe)}

\item{penalty.factor.lasso}{factors applied to each coefficient in lasso penalty}

\item{penalty.factor.trlasso}{factors applied to each coefficient in trlasso penalty}

\item{penalty.factor.thresh}{lower limit for penalty factors (default 1e-2)}

\item{...}{parameters of trlasso function}
}
\value{
lasso model
\item{fit}{lasso model with hole data}
\item{lambda.min}{lambda with minimum cross validation error}
\item{lambda.min.index}{index of lambda.min}
\item{lambda.1se}{largest lambda such that error is within 1 standard error of the minimum}
\item{lambda.1se.index}{index of lambda.1se}
\item{foldid}{fold id}
\item{cve}{cross validation error}
\item{cvse}{cross validation standard error}
\item{cvup}{cross validation error + standard error}
\item{cvlo}{cross validation error - standard error}
\item{fit.preval}{array of prevalidated fits}
}
\description{
Fit a model using a design matrix with cross validation
}
\examples{
X <- as.matrix(iris[, 1:3])
y <- iris[, 4]
cv_fit <- cv.trlasso(X, y, nlambda=50, lambda.min.ratio=1e-2)
plot(cv_fit)
plot(cv_fit$fit)

}
