% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predict.trlasso}
\alias{predict.trlasso}
\title{Predict responses}
\usage{
\method{predict}{trlasso}(object, newx, s = NULL, type = "response", ...)
}
\arguments{
\item{object}{trlasso model}

\item{newx}{matrix of explanatory variables}

\item{s}{selected lambda (default: all)}

\item{type}{of prediction for logistic regression: "response"(default) or "class"}

\item{...}{parameters of predict function}
}
\description{
Predict responses
}
\examples{
X <- as.matrix(iris[, 1:3])
y <- iris[, 4]
cv_fit <- cv.trlasso(X, y, nlambda=50, lambda.min.ratio=1e-2)
predict(cv_fit$fit, X)

}
