% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{softThresholdC2}
\alias{softThresholdC2}
\title{soft thresholding function for transfer lasso}
\usage{
softThresholdC2(z, g1, g2, b)
}
\arguments{
\item{z}{z}

\item{g1}{gamma1}

\item{g2}{gamma2}

\item{b}{b}
}
\value{
value
}
\description{
soft thresholding function for transfer lasso
}
\keyword{internal}
