% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{trlasso}
\alias{trlasso}
\title{Fit a model using a design matrix}
\usage{
trlasso(
  X,
  y,
  beta_tilde = NULL,
  family = "gaussian",
  lambda.min.ratio = 0.01,
  nlambda = 100,
  lambda = NULL,
  alpha = 1,
  eps = 1e-04,
  verbose = FALSE,
  standardize = TRUE,
  penalty.factor.lasso = NULL,
  penalty.factor.trlasso = NULL,
  penalty.factor.thresh = 0.01,
  ...
)
}
\arguments{
\item{X}{matrix of explanatory variables}

\item{y}{vector of objective variable}

\item{beta_tilde}{vector of initial parameters}

\item{family}{family of regression: "gaussian" is only supported}

\item{lambda.min.ratio}{ratio of max lambda and min lambda (ignored if lambda is specified)}

\item{nlambda}{the number of lambda (ignored if lambda is specified)}

\item{lambda}{lambda sequence}

\item{alpha}{alpha}

\item{eps}{convergence threshold (default 1e-4)}

\item{verbose}{whether output verbose warnings and messages (default FALSE)}

\item{standardize}{standardize or not (experimental, not safe)}

\item{penalty.factor.lasso}{factors applied to each coefficient in lasso penalty}

\item{penalty.factor.trlasso}{factors applied to each coefficient in trlasso penalty}

\item{penalty.factor.thresh}{lower limit for penalty factors (default 1e-2)}

\item{...}{parameters for optimization}
}
\value{
lasso model
\item{beta}{coefficients}
\item{beta_standard}{standardized coefficients}
\item{a0}{intercepts}
\item{lambda}{regularization parameters}
\item{family}{family}
}
\description{
Fit a model using a design matrix
}
\examples{
X <- as.matrix(iris[, 1:3])
y <- iris[, 4]
fit <- trlasso(X, y, nlambda=50, lambda.min.ratio=1e-2)
plot(fit)

}
